pyproject.toml
[tool.poetry]
name = "django-api-dev"
version = "0.1.0"
description = ""
authors = ["Admin <Dexter2366@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
psycopg2 = "^2.9.9"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
pytest-django = "^4.7.0"
pytest-factoryboy = "^2.6.0"
pillow = "^10.1.0"
django-cors-headers = "^4.3.1"
python-dotenv = "^1.0.0"
drf-spectacular = "^0.26.5"
djoser = "^2.2.2"
django-filter = "^23.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
